tests/wpt/tests/css/css-text/parsing/text-spacing-trim-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-spacing-trim with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-spacing-trim-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-spacing-trim", "normal");
test_valid_value("text-spacing-trim", "auto");
test_valid_value("text-spacing-trim", "space-all");
test_valid_value("text-spacing-trim", "trim-auto");
test_valid_value("text-spacing-trim", "trim-all");
test_valid_value("text-spacing-trim", "trim-start");
test_valid_value("text-spacing-trim", "space-first");
</script>
</body>
</html>
